# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrisisMesh Emergency Response API",
    "description": "Gateways 2026 Domain 4: Crisis Command - Multi-Agent Coordination Control Room API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticate and obtain a development JWT token with embedded RBAC role.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Current User Profile",
        "description": "Retrieve identity and authorized role of the authenticated token bearer.",
        "operationId": "get_current_user_profile_auth_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports": {
      "post": {
        "tags": [
          "Operations & Ingestion"
        ],
        "summary": "Submit Report",
        "description": "Submit raw citizen or agency report. Requires OPERATOR or COMMANDER role.",
        "operationId": "submit_report_reports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors": {
      "post": {
        "tags": [
          "Operations & Ingestion"
        ],
        "summary": "Submit Sensor Reading",
        "description": "Ingest physical telemetry (water gauges, rain meters). Requires valid HMAC-SHA256 signature.",
        "operationId": "submit_sensor_reading_sensors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/units/{unit_id}/status": {
      "post": {
        "tags": [
          "Operations & Ingestion"
        ],
        "summary": "Update Unit Status",
        "description": "Update emergency unit operational state (idle, en_route, unavailable). Triggers selective re-plan.",
        "operationId": "update_unit_status_units__unit_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roads/block": {
      "post": {
        "tags": [
          "Operations & Ingestion"
        ],
        "summary": "Block Road",
        "description": "Dynamically declare a road segment (u, v) or geographical zone blocked by floods.",
        "operationId": "block_road_roads_block_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadBlockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/audit": {
      "get": {
        "tags": [
          "Audit & Security"
        ],
        "summary": "Get Audit Trail",
        "description": "Retrieve full append-only tamper-evident audit chain.",
        "operationId": "get_audit_trail_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit/verify": {
      "get": {
        "tags": [
          "Audit & Security"
        ],
        "summary": "Verify Audit Integrity",
        "description": "Perform mathematical verification of SHA-256 hash chaining from Genesis to head.",
        "operationId": "verify_audit_integrity_audit_verify_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/security/events": {
      "get": {
        "tags": [
          "Audit & Security"
        ],
        "summary": "Get Security Events",
        "description": "Retrieve all security violations, injection attempts, and perimeter alerts.",
        "operationId": "get_security_events_security_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "ReportIngestRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "ip_address": {
            "type": "string",
            "title": "Ip Address",
            "default": "127.0.0.1"
          }
        },
        "type": "object",
        "required": [
          "text",
          "source_id",
          "lat",
          "lon"
        ],
        "title": "ReportIngestRequest"
      },
      "ReportIngestResponse": {
        "properties": {
          "report_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_quarantined": {
            "type": "boolean",
            "title": "Is Quarantined"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "security_events": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Security Events"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "status",
          "is_quarantined",
          "confidence"
        ],
        "title": "ReportIngestResponse"
      },
      "RoadBlockRequest": {
        "properties": {
          "u": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "U"
          },
          "v": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "V"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "radius_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius Km"
          }
        },
        "type": "object",
        "title": "RoadBlockRequest"
      },
      "Role": {
        "type": "string",
        "enum": [
          "viewer",
          "operator",
          "commander"
        ],
        "title": "Role"
      },
      "SensorIngestRequest": {
        "properties": {
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id"
          },
          "sensor_type": {
            "type": "string",
            "title": "Sensor Type",
            "default": "water_level"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "meters"
          },
          "flood_threshold": {
            "type": "number",
            "title": "Flood Threshold",
            "default": 0.5
          },
          "signature_hex": {
            "type": "string",
            "title": "Signature Hex"
          }
        },
        "type": "object",
        "required": [
          "sensor_id",
          "lat",
          "lon",
          "value",
          "signature_hex"
        ],
        "title": "SensorIngestRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "public_key_hex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Key Hex"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds",
            "default": 86400
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role",
          "user_id"
        ],
        "title": "TokenResponse"
      },
      "UnitStatusUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Operational state: idle, en_route, unavailable"
          },
          "current_target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Target Id"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UnitStatusUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}